render a list of shops
