+++ src/shops/components/shop/Shop.scss
@@ -1,0 +1,5 @@
+.shop {
+  padding: 1em;
+  margin: 1em;
+  box-shadow: 1px 1px 1px #dedede;
+}
